requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib

qdrant-client
llama-index
llama-index-embeddings-openai-like
llama-index-llms-openai-like
scikit-learn         
joblib

nltk

fastapi[all]
uvicorn[standard]

pydantic
pydantic-settings    
python-dotenv